requirements.txt
streamlit
weaviate-client
openai
python-dotenv
pytesseract
Pillow
pdf2image
PyMuPDF
opencv-python
moviepy
spacy
keybert
requests
pandas
python-pptx
xlrd
beautifulsoup4
scikit-learn
nltk
plotly